refactor(http-benchmarks): move scenario setups to a scenario-setups folder
