+++ packages/http/tools/http-benchmarks/src/scenario-setups/currentInversifyExpress/setupBasicGetScenario.ts
