+++ packages/http/tools/http-benchmarks/src/scenario-setups/express/setupBasicGetScenario.ts
