+++ packages/http/tools/http-benchmarks/src/scenario-setups/fastify/setupBasicGetScenario.ts
