+++ packages/http/tools/http-benchmarks/src/scenario-setups/nestJsExpress/setupBasicGetScenario.ts
@@ -1,7 +1,7 @@
 import { NestApplication, NestFactory } from '@nestjs/core';
 
 import { DEFAULT_PORT } from '../../constant/defaultPort';
-import { BasicGetAppModule } from '../nestJS/BasicGetAppModule';
+import { BasicGetAppModule } from '../../scenario/nestJs/BasicGetAppModule';
 
 async function setUp(): Promise<void> {
   const app: NestApplication = await NestFactory.create(BasicGetAppModule, {
