+++ packages/http/tools/http-benchmarks/src/scenario-setups/nestJsFastify/setupBasicGetScenario.ts
@@ -2,7 +2,7 @@ import { NestApplication, NestFactory } from '@nestjs/core';
 import { FastifyAdapter } from '@nestjs/platform-fastify';
 
 import { DEFAULT_PORT } from '../../constant/defaultPort';
-import { BasicGetAppModule } from '../nestJS/BasicGetAppModule';
+import { BasicGetAppModule } from '../../scenario/nestJs/BasicGetAppModule';
 
 async function setUp(): Promise<void> {
   const app: NestApplication = await NestFactory.create(
