+++ packages/http/tools/http-benchmarks/src/scenario/fastify/FastifyBasicGetScenario.ts
@@ -3,6 +3,10 @@ import { Platform } from '../models/Platform';
 
 export class FastifyBasicGetScenario extends BaseK6Scenario {
   constructor() {
-    super(Platform.fastify, 'BasicGetScenario.ts', 'fastify/setup.js');
+    super(
+      Platform.fastify,
+      'BasicGetScenario.ts',
+      'fastify/setupBasicGetScenario.js',
+    );
   }
 }
